requirements.txt
streamlit>=1.20
google-generativeai>=0.22.0
gspread>=5.7.0
oauth2client>=4.1.3
python-dotenv>=1.0
